# api.yaml
swagger: "2.0"
info:
  title: Social Media API
  description: API documentation
  version: 1.0.0
paths:
...
  /refresh:
    get:
      summary: Get new AccessToken
      operationId: RefreshToken
      tags: 
        - Users
      security:
        - cookieAuth: []
      responses:
        200:
          description: OK
...
  /users:
    get:
      summary: Get all users
      operationId: AllUsers
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
  /users/{id}:
    get:
      summary: Get a user
      operationId: User
      tags:
        - Users
      parameters:
        - $ref: "#/parameters/idParams"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK

    delete:
      summary: Delete a user
      operationId: DeleteUser
      tags:
        - Users
      parameters:
        - $ref: "#/parameters/idParams"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
  /users/updateMe:
    put:
      summary: Update your username and re-login to get a new valid accessToken
      operationId: UpdateUser
      tags:
        - Users
      parameters:
        - in: body
          name: username
          required: true
          schema:
            $ref: "#/definitions/Username"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
  /users/updateMyPassword:
    put:
      summary: Update User Password
      operationId: UpdatePassword
      tags:
        - Users
      parameters:
        - in: body
          name: passwordUpdate
          required: true
          schema:
            $ref: "#/definitions/Password"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
  /posts:
    get:
      summary: Get all posts
      operationId: AllPosts
      tags:
        - Posts
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK

    post:
      summary: Create a post
      operationId: CreatePost
      tags:
        - Posts
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: text
          description: Write Your Post Here
          type: string
          required: true
        - in: formData
          name: photo
          type: file
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
  /posts/{id}:
    get:
      summary: Get a Post
      operationId: SinglePost
      tags:
        - Posts
      parameters:
        - $ref: "#/parameters/postParams"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK

    delete:
      summary: Delete a Post
      operationId: DeletePost
      tags:
        - Posts
      parameters:
        - $ref: "#/parameters/postParams"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
...
  /posts/comments/{id}:
    get:
      summary: Get Post's comments
      operationId: PostComments
      tags:
        - Posts
      parameters:
        - $ref: "#/parameters/postParams"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
  /posts/post/{id}:
    get:
      summary: Get Total comments on a Post
      operationId: TotalComments
      tags: [Posts]
      parameters:
        - $ref: "#/parameters/postParams"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
...
  /following:
    post:
      summary: Follow a User
      operationId: FollowUser
      tags:
        - Users
      parameters:
        - in: body
          name: User ID
          required: true
          schema:
            $ref: "#/definitions/Following"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK

    get:
      summary: View your following status
      operationId: Following
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
  /following/{id}:
    get:
      summary: View a User's following status
      operationId: UserFollowing
      tags: [Users]
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: The User's ID
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK

    delete:
      summary: Unfollow a User
      operationId: UnfollowUser
      tags: [Users]
      parameters:
        - $ref: "#/parameters/followingParams"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt 
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
parameters:
  idParams:
    in: path
    name: id
    required: true
    type: string
    minimum: 1
    description: The User ID
  postParams:
    in: path
    name: id
    required: true
    type: string
    minimum: 1
    description: The Post's ID
...
  followingParams:
    in: path
    name: id
    required: true
    type: string
    minimum: 1
    description: ID of the user you want to unfollow
...
definitions:
...
  Following:
    type: object
    properties:
      followingId:
        type: string
        example: The ID of the User you want to follow
    required:
      - followingId
...
  Username:
    type: object
    properties:
      username:
        type: string
    required: [username]
  Password:
    type: object
    properties:
      passwordCurrent:
        type: string
      password:
        type: string
        minLength: 8
      passwordConfirm:
        type: string
        minLength: 8
    required: [passwordCurrent, password, passwordConfirm]
...